spanreceiver: define SpanReceiver interface
An interface `SpanReceiver` that receives:
    (*commonpb.Node, ...*tracepb.Span)

to uniquely identify a data source by node alongside
the spans it has received from that source.
It sends back an acknowledgement and an error.

The return signature is `(*Acknowledgement, error)` because it
would useful to return to callers information about how many spans
were successful, errors that were encountered, allowing for batches
of spans to be written like we would with `io.Write`. This will allow
callers to retry, report success and failure rates too.

Fixes #30