Increase resource limits for some perf tests (#1134)
A few perf tests are recently failing once in a while on CircleCI.
I checked the commit history and was unable to identify a commit
that may have caused a regression. The failures are happening
randomly and the only conclusion I can make is that CircleCI is
a bit slower than usual and these tests have limits that are too
tight.

I am increasing the limits to avoid the intermittent CI failures.