Do not embed ProtocolSettings in gRPC (#1210)
Adding `ProtocolServerSettings` was a bit of a rush. We determined that `endpoint` has different meaning based on protocol, also not all the protocols support TLS.

In this PR we revert embedding `ProtocolServerSettings` in the GRPCServerSettings and make it consistent with HttpServerSettings.

Work left: Consistent config name for TLS settings `tls_settings` or `tls_credentials`.